feat: Add driver_id parameter to insertDelivery function
The code changes in `insertDelivery.ts` modify the `insertDelivery` function to accept a `driver_id` parameter. This allows for more flexibility in assigning a driver to a delivery. The function now inserts the `driver_id` value into the database when creating a new delivery. This enhancement improves the functionality of the application.